# asg_form/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ASG 赛事官网-后端API文档",
    "description": "这是由罗澜使用ASP.NET.Core开发的ASG赛事组后端系统，包括官网和后台管理系统。使用 sqlserver作为数据库，identity框架进行账号控制。",
    "version": "V 1.9.6"
  },
  "paths": {
...
    "/api/v1/admin/game": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "发布一个竞猜比赛",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "schema": {
              "$ref": "#/definitions/req_team_game"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "schedule"
        ],
        "summary": "删除竞猜比赛",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "gameid",
            "description": "",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/admin/game/win": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "发布胜利者",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "teamid",
            "description": "",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "winteam",
            "description": "",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/api/v1/game/pushgame": {
      "put": {
        "tags": [
          "schedule"
        ],
        "summary": "投票",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "gameid",
            "description": "",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "teamid",
            "description": "",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "req_team_game": {
      "type": "object",
      "properties": {
        "team1_name": {
          "type": "string"
        },
        "team2_name": {
          "type": "string"
        },
        "opentime": {
          "format": "date-time",
          "type": "string"
        },
        "commentary": {
          "description": "解说的名字，用逗号隔开！！！！！！！",
          "type": "string"
        },
        "referee": {
          "description": "裁判的名字",
          "type": "string"
        }
      }
    },
...
  }
}